+++ requirements.txt
@@ -23,9 +23,6 @@ joblib==1.0.1
 jupyter-client==6.1.12
 jupyter-core==4.7.1
 MarkupSafe==1.1.1
-mkl-fft==1.3.0
-mkl-random==1.2.1
-mkl-service==2.3.0
 numpy==1.20.1
 olefile==0.46
 packaging==20.9
